Attaque avec les fleches plutot que la souris, dash avec espace

+++ include/attributs_personnages.h
@@ -3,7 +3,7 @@ int pv; /**< Points de vie */
 
 e_nom_attaque nom_attaque; /**< Nom de l'attaque équipée */
 t_attaque_tir attaque_tir_equipee; /**< Descriptif de l'attaque équipée */
-int attaque_vx; /**< Vecteur direction de l'attaque en x */
-int attaque_vy; /**< Vecteur direction de l'attaque en y */
+float attaque_vx; /**< Vecteur direction de l'attaque en x */
+float attaque_vy; /**< Vecteur direction de l'attaque en y */
 
 //t_attaque_corps attaque_corps_equipee;
